replaced scroll with toggle
