+++ src/components/Audioplayer/ScrollButton/index.js
@@ -1,32 +1,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import LocaleFormattedMessage from 'components/LocaleFormattedMessage';
 
-import { StyledPopover } from '../RepeatDropdown/index';
+import SwitchToggle from 'quran-components/lib/Toggle';
 
-// eslint-disable-next-line no-confusing-arrow
-const scrollColor = props =>
-  props.active ? props.theme.brandPrimary : props.theme.textColor;
-
-const ScrollLink = styled.a`
-  width: 100%;
-  display: inline-block;
-  cursor: pointer;
-  padding-right: 1.5%;
-  color: ${scrollColor};
-  outline: none;
-  cursor: pointer;
-  margin-bottom: 0;
-`;
+import { ControlButton, StyledPopover } from '../index';
 
 const ScrollButton = ({ shouldScroll, onScrollToggle }) => {
   const tooltip = (
-    <StyledPopover id="scroll-popover">
-      <LocaleFormattedMessage
-        id="player.scrollButtonTip"
-        defaultMessage="Automatically scrolls to the currently playing ayah on transitions..."
+    <StyledPopover
+      id="scroll-toggle-popoverr"
+      title={
+        <LocaleFormattedMessage
+          id="player.scrollButtonTip"
+          defaultMessage="Automatically scrolls to the currently playing ayah on transitions..."
+        />
+      }
+    >
+      {'  '}
+      <SwitchToggle
+        checked={shouldScroll}
+        onToggle={onScrollToggle}
+        id="scroll-toggle"
+        name="scroll-toggle"
+        flat
       />
     </StyledPopover>
   );
@@ -36,16 +34,12 @@ const ScrollButton = ({ shouldScroll, onScrollToggle }) => {
       <OverlayTrigger
         overlay={tooltip}
         placement="top"
-        trigger={['hover', 'focus']}
-        delayHide={150}
+        trigger="click"
+        rootClose
       >
-        <ScrollLink
-          tabIndex="-1"
-          active={shouldScroll}
-          onClick={onScrollToggle}
-        >
+        <ControlButton active={shouldScroll}>
           <i className="ss-icon ss-link" />
-        </ScrollLink>
+        </ControlButton>
       </OverlayTrigger>
     </div>
   );
